Add UInt32+uppercase file & Edit private variable addressesGroup
